+++ ERPServer/Shared/Shared/Entities/Setup/ProductDTO.cs
@@ -15,9 +15,9 @@ namespace Shared.Entities.Setup
         public decimal Price { get; set; }
         public decimal SellingPricePercentage { get; set; }//سعر البيع
         public decimal PurchasingPricePercentage { get; set; }//سعر الشراء
-        public int ActualQuantity { get; set; }
-        public int LowQuantity { get; set; }
-        public int HighQuantity { get; set; }
+        public decimal ActualQuantity { get; set; }
+        public decimal LowQuantity { get; set; }
+        public decimal HighQuantity { get; set; }
         public long CategoryId { get; set; }
         public string CategoryName { get; set; }
         public string ImageBase64 { get; set; }
